chess/requirements.txt
flask
gunicorn
matplotlib
mpld3
ndjson
pytz
python-dateutil
requests
